full version of load and little changes in derivatives of Cargo

+++ shm/include/item.hpp
@@ -9,7 +9,7 @@ public:
     Item(const std::string& name, size_t amount, size_t basePrice, Rarity rarity);
 
     // override from Cargo
-    size_t getPrice() const override { return basePrice_ * static_cast<int>(rarity_); }
+    size_t getPrice() const override { return basePrice_ * static_cast<size_t>(rarity_); }
     std::string getName() const override { return name_; }
     size_t getAmount() const override { return amount_; }
     size_t getBasePrice() const override { return basePrice_; }
